requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.3
pymongo==4.6.0
playwright==1.40.0
apscheduler==3.10.4
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
reportlab==4.0.7
pdfkit==1.0.0
google-generativeai==0.3.0
groq==0.4.2
huggingface-hub==0.19.4
openrouter==0.0.22
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
tenacity==8.2.3
